examples/mlflow_plugin/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
absl-py==1.4.0
    # via
    #   tensorboard
    #   tensorflow
alembic==1.11.1
    # via mlflow
arrow==1.2.3
    # via jinja2-time
astunparse==1.6.3
    # via tensorflow
binaryornot==0.4.4
    # via cookiecutter
blinker==1.6.2
    # via flask
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via
    #   kubernetes
    #   requests
cffi==1.15.1
    # via cryptography
chardet==5.1.0
    # via binaryornot
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   cookiecutter
    #   databricks-cli
    #   flask
    #   nebulakit
    #   mlflow
cloudpickle==2.2.1
    # via
    #   nebulakit
    #   mlflow
contourpy==1.1.0
    # via matplotlib
cookiecutter==2.1.1
    # via nebulakit
croniter==1.4.1
    # via nebulakit
cryptography==41.0.1
    # via
    #   pyopenssl
    #   secretstorage
cycler==0.11.0
    # via matplotlib
databricks-cli==0.17.7
    # via mlflow
dataclasses-json==0.5.9
    # via nebulakit
decorator==5.1.1
    # via retry
deprecated==1.2.14
    # via nebulakit
diskcache==5.6.1
    # via nebulakit
docker==6.1.3
    # via
    #   nebulakit
    #   mlflow
docker-image-py==0.1.12
    # via nebulakit
docstring-parser==0.15
    # via nebulakit
entrypoints==0.4
    # via mlflow
flask==2.3.2
    # via mlflow
flatbuffers==1.12
    # via tensorflow
nebulaidl==1.2.11
    # via nebulakit
nebulakit==1.2.11
    # via nebulakitplugins-mlflow
nebulakitplugins-mlflow==1.7.0
    # via -r requirements.in
fonttools==4.40.0
    # via matplotlib
gast==0.4.0
    # via tensorflow
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via
    #   nebulakit
    #   mlflow
google-auth==2.21.0
    # via
    #   google-auth-oauthlib
    #   kubernetes
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
googleapis-common-protos==1.56.4
    # via
    #   nebulaidl
    #   grpcio-status
greenlet==2.0.2
    # via sqlalchemy
grpcio==1.48.2
    # via
    #   nebulakit
    #   grpcio-status
    #   tensorboard
    #   tensorflow
grpcio-status==1.48.2
    # via nebulakit
gunicorn==20.1.0
    # via mlflow
h5py==3.9.0
    # via tensorflow
idna==3.4
    # via requests
importlib-metadata==6.7.0
    # via
    #   alembic
    #   flask
    #   nebulakit
    #   keyring
    #   markdown
    #   mlflow
importlib-resources==5.12.0
    # via
    #   alembic
    #   keyring
    #   matplotlib
itsdangerous==2.1.2
    # via flask
jaraco-classes==3.2.3
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.2
    # via
    #   cookiecutter
    #   flask
    #   jinja2-time
    #   mlflow
jinja2-time==0.2.0
    # via cookiecutter
joblib==1.3.1
    # via
    #   nebulakit
    #   scikit-learn
keras==2.9.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
keyring==24.2.0
    # via nebulakit
kiwisolver==1.4.4
    # via matplotlib
kubernetes==26.1.0
    # via nebulakit
libclang==16.0.0
    # via tensorflow
mako==1.2.4
    # via alembic
markdown==3.4.3
    # via
    #   mlflow
    #   tensorboard
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
    #   marshmallow-jsonschema
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow-jsonschema==0.13.0
    # via nebulakit
matplotlib==3.7.2
    # via mlflow
mlflow==2.8.1
    # via nebulakitplugins-mlflow
more-itertools==9.1.0
    # via jaraco-classes
mypy-extensions==1.0.0
    # via typing-inspect
natsort==8.4.0
    # via nebulakit
numpy==1.23.5
    # via
    #   contourpy
    #   nebulakit
    #   h5py
    #   keras-preprocessing
    #   matplotlib
    #   mlflow
    #   opt-einsum
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
oauthlib==3.2.2
    # via
    #   databricks-cli
    #   requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==23.1
    # via
    #   docker
    #   marshmallow
    #   matplotlib
    #   mlflow
    #   plotly
    #   tensorflow
pandas==1.5.3
    # via
    #   nebulakit
    #   mlflow
pillow==10.0.0
    # via matplotlib
plotly==5.15.0
    # via nebulakitplugins-mlflow
protobuf==3.20.1
    # via
    #   -r requirements.in
    #   nebulaidl
    #   nebulakit
    #   googleapis-common-protos
    #   grpcio-status
    #   mlflow
    #   protoc-gen-swagger
    #   tensorboard
    #   tensorflow
protoc-gen-swagger==0.1.0
    # via nebulaidl
py==1.11.0
    # via retry
pyarrow==10.0.1
    # via
    #   nebulakit
    #   mlflow
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt==2.7.0
    # via databricks-cli
pyopenssl==23.2.0
    # via nebulakit
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   arrow
    #   croniter
    #   nebulakit
    #   kubernetes
    #   matplotlib
    #   pandas
python-json-logger==2.0.7
    # via nebulakit
python-slugify==8.0.1
    # via cookiecutter
pytimeparse==1.1.8
    # via nebulakit
pytz==2023.3
    # via
    #   nebulakit
    #   mlflow
    #   pandas
pyyaml==6.0
    # via
    #   cookiecutter
    #   nebulakit
    #   kubernetes
    #   mlflow
    #   responses
querystring-parser==1.2.4
    # via mlflow
regex==2023.6.3
    # via docker-image-py
requests==2.31.0
    # via
    #   cookiecutter
    #   databricks-cli
    #   docker
    #   nebulakit
    #   kubernetes
    #   mlflow
    #   requests-oauthlib
    #   responses
    #   tensorboard
requests-oauthlib==1.3.1
    # via
    #   google-auth-oauthlib
    #   kubernetes
responses==0.23.1
    # via nebulakit
retry==0.9.2
    # via nebulakit
rsa==4.9
    # via google-auth
scikit-learn==1.3.0
    # via mlflow
scipy==1.10.1
    # via
    #   mlflow
    #   scikit-learn
secretstorage==3.3.3
    # via keyring
six==1.16.0
    # via
    #   astunparse
    #   databricks-cli
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   kubernetes
    #   python-dateutil
    #   querystring-parser
    #   tensorflow
smmap==5.0.0
    # via gitdb
sortedcontainers==2.4.0
    # via nebulakit
sqlalchemy==2.0.18
    # via
    #   alembic
    #   mlflow
sqlparse==0.4.4
    # via mlflow
statsd==3.3.0
    # via nebulakit
tabulate==0.9.0
    # via databricks-cli
tenacity==8.2.2
    # via plotly
tensorboard==2.9.0
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.9.0
    # via -r requirements.in
tensorflow-estimator==2.9.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.32.0
    # via tensorflow
termcolor==2.3.0
    # via tensorflow
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.1.0
    # via scikit-learn
types-pyyaml==6.0.12.10
    # via responses
typing-extensions==4.7.1
    # via
    #   alembic
    #   nebulakit
    #   sqlalchemy
    #   tensorflow
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==1.26.16
    # via
    #   databricks-cli
    #   docker
    #   nebulakit
    #   google-auth
    #   kubernetes
    #   requests
    #   responses
websocket-client==1.6.1
    # via
    #   docker
    #   kubernetes
werkzeug==2.3.8
    # via
    #   flask
    #   tensorboard
wheel==0.40.0
    # via
    #   astunparse
    #   nebulakit
    #   tensorboard
wrapt==1.15.0
    # via
    #   deprecated
    #   nebulakit
    #   tensorflow
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
